WebFrameDocs/src/demo/AuthSystem/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.10.0
django-redis==5.0.0
PyJWT==2.1.0
pytz==2021.3
redis==3.5.3
sqlparse==0.4.2
